observability_security/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
PyJWT
Werkzeug
python-dotenv
jwt 
requests
argon2-cffi
flask-jwt-extended